refactor(Server): Update service-orders domain to satisfy new repo architecture
