--- packages/db/src/schema/device-categories.ts
@@ -1,12 +1,0 @@
-import { createId } from "@paralleldrive/cuid2";
-import { mysqlTable, varchar } from "drizzle-orm/mysql-core";
-import { createSelectSchema } from "drizzle-zod";
-
-export const deviceCategories = mysqlTable("device_categories", {
-	id: varchar("id", { length: 25 })
-		.$defaultFn(() => createId())
-		.primaryKey(),
-	name: varchar("name", { length: 100 }).notNull(),
-});
-
-export const deviceCategorySelectSchema = createSelectSchema(deviceCategories);
