+++ packages/db/src/schema/index.ts
@@ -1,8 +1,8 @@
 export * from "./clients";
 export * from "./companies";
-export * from "./device-brands";
-export * from "./device-categories";
 export * from "./employees";
+export * from "./model-categories";
+export * from "./model-makers";
 export * from "./one-time-tokens";
 export * from "./refresh-tokens";
 export * from "./service-order-photos";
