+++ packages/db/src/schema/service-orders.ts
@@ -10,9 +10,9 @@ import { createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
 import { clients } from "./clients";
 import { companies } from "./companies";
-import { deviceBrands } from "./device-brands";
-import { deviceCategories } from "./device-categories";
 import { employees } from "./employees";
+import { modelCategories } from "./model-categories";
+import { modelMakers } from "./model-makers";
 
 export const serviceOrderStatusEnum = mysqlEnum("status", [
 	"pending",
@@ -37,11 +37,11 @@ export const serviceOrders = mysqlTable("service_orders", {
 	employeeId: varchar("employee_id", { length: 25 })
 		.references(() => employees.id, { onDelete: "restrict" })
 		.notNull(),
-	deviceBrandId: varchar("device_brand_id", { length: 25 })
-		.references(() => deviceBrands.id, { onDelete: "restrict" })
+	deviceMakerId: varchar("device_brand_id", { length: 25 })
+		.references(() => modelMakers.id, { onDelete: "restrict" })
 		.notNull(),
 	deviceCategoryId: varchar("device_category_id", { length: 25 })
-		.references(() => deviceCategories.id, { onDelete: "restrict" })
+		.references(() => modelCategories.id, { onDelete: "restrict" })
 		.notNull(),
 	deviceModel: varchar("device_model", { length: 100 }).notNull(),
 	imei: varchar("imei", { length: 50 }),
